PatientDocumentGenerator/requirements.txt
azure-common==1.1.24
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
idna==2.9
lxml==4.5.0
pycparser==2.19
python-dateutil==2.8.1
python-docx==0.8.10
requests==2.23.0
six==1.14.0
urllib3==1.25.8
